config_system: move expression handling
Move expression handling code into expr.py
