+++ config_system/config_system/__init__.py
@@ -26,7 +26,6 @@ if os.path.isfile("build/make/core/envsetup.mk"):
 
 from config_system.general import \
     can_enable, \
-    format_dependency_list, \
     get_config_bool, \
     get_config_int, \
     get_config_string, \
@@ -40,3 +39,6 @@ from config_system.data import \
     get_config, \
     get_config_list, \
     get_mconfig_dir  # nopep8: E402 module level import not at top of file
+
+from config_system.expr import \
+    format_dependency_list  # nopep8: E402 module level import not at top of file
