+++ config_system/config_system/general.py
@@ -23,215 +23,18 @@ logger = logging.getLogger(__name__)
 logger.addHandler(logging.NullHandler())
 
 
-def check_depends(depends, value):
-    """Check if the config identified in value is a simple dependency
-    listed in the depends expression.
-    A simple expression consists of just && and || boolean operators.
-    If the expression uses any other operator, return False.
-
-    This is used by the menu_parse below to indent dependent configs.
-    """
-    if depends is None:
-        return False
-    assert type(depends) == tuple
-    assert len(depends) in [2, 3]
-
-    if depends[0] == 'and':
-        return (check_depends(depends[1], value) or
-                check_depends(depends[2], value))
-    elif depends[0] == 'or':
-        return (check_depends(depends[1], value) and
-                check_depends(depends[2], value))
-    elif depends[0] == 'word':
-        return depends[1] == value
-    return False
-
-
-# Operators where the result is not boolean
-ARITH_SET = {"+", "-"}
-
-
-def _expr_value(e):
-    """Evaluate the value of the input expression.
-    This isn't expected to use conditional operators.
-    """
-    assert type(e) == tuple
-    assert len(e) in [2, 3]
-    if len(e) == 3:
-        left = _expr_value(e[1])
-        right = _expr_value(e[2])
-        if type(left) != type(right):
-            raise TypeError("'{}' operator is not valid with mixed types".format(e[0]))
-        elif left in ['y', 'n'] or right in ['y', 'n']:
-            raise TypeError("'{}' operator is not valid on booleans".format(e[0]))
-        elif e[0] == '+':
-            return left + right
-        elif e[0] == '-':
-            if type(left) is str:
-                raise TypeError("'-' operator is not valid on strings")
-                return left
-            return left - right
-    elif e[0] == 'string':
-        return e[1]
-    elif e[0] == 'number':
-        return e[1]
-    elif e[0] == 'boolean':
-        return e[1]
-    elif e[0] == 'word':
-        return data.get_config(e[1])['value']
-
-    raise Exception("Unexpected depend list: " + str(e))
-
-
-def expr_value(e):
-    try:
-        result = _expr_value(e)
-    except TypeError as err:
-        logging.error("{} in expression '{}'".format(str(err), format_dependency_list(e)))
-        result = ""
-
-    return result
-
-
-def _condexpr_value(e):
-    """Evaluate the value of the input expression.
-    Note that booleans are propagated as 'y' and 'n'
-    """
-    assert type(e) == tuple
-    assert len(e) in [2, 3]
-
-    if len(e) == 3:
-        if e[0] in ARITH_SET:
-            return _expr_value(e)
-
-        left = _condexpr_value(e[1])
-        right = _condexpr_value(e[2])
-        if type(left) != type(right):
-            # Boolean result expected
-            return 'n'
-        elif e[0] == 'and':
-            if left == 'y' and right == 'y':
-                return 'y'
-            return 'n'
-        elif e[0] == 'or':
-            if left == 'y' or right == 'y':
-                return 'y'
-            return 'n'
-        elif e[0] == '=':
-            if left == right:
-                return 'y'
-            return 'n'
-        elif e[0] == '!=':
-            if left != right:
-                return 'y'
-            return 'n'
-        elif e[0] == '>':
-            if left > right:
-                return 'y'
-            return 'n'
-        elif e[0] == '>=':
-            if left >= right:
-                return 'y'
-            return 'n'
-        elif e[0] == '<':
-            if left < right:
-                return 'y'
-            return 'n'
-        elif e[0] == '<=':
-            if left <= right:
-                return 'y'
-            return 'n'
-    elif e[0] == 'not':
-        if _condexpr_value(e[1]) == 'y':
-            return 'n'
-        return 'y'
-    elif e[0] == 'string':
-        return e[1]
-    elif e[0] == 'number':
-        return e[1]
-    elif e[0] == 'boolean':
-        return e[1]
-    elif e[0] == 'word':
-        return data.get_config(e[1])['value']
-
-    raise Exception("Unexpected depend list: " + str(e))
-
-
-def condexpr_value(e):
-    if e is None:
-        return 'y'
-    try:
-        result = _condexpr_value(e)
-    except TypeError as err:
-        logging.error("{} in expression '{}'".format(str(err), format_dependency_list(e)))
-        result = 'n'
-
-    return result
-
-
 def can_enable(depends):
-    value = condexpr_value(depends)
+    value = expr.condexpr_value(depends)
     if value == 'y':
         return True
     return False
 
 
 def is_visible(cond):
-    value = condexpr_value(cond)
+    value = expr.condexpr_value(cond)
     if value == 'n':
         return False
     return True
-
-
-def dependency_list(e):
-    """
-    Get the set of config identifiers referred to by an expression. A
-    set is returned instead of a list as we don't need duplicates, and
-    order doesn't matter.
-    """
-    if e is None:
-        return set()
-    assert type(e) == tuple
-
-    if e[0] in ['and', 'or', '=', '!=', '<', '<=', '>', '>=', '+', '-']:
-        return dependency_list(e[1]) | dependency_list(e[2])
-    elif e[0] == 'not':
-        return dependency_list(e[1])
-    elif e[0] in ['string', 'number', 'boolean']:
-        # Quoted string, number or boolean
-        return set()
-    elif e[0] == 'word':
-        return {e[1]}
-    raise Exception("Unexpected depend list: " + str(e))
-
-
-OPERATOR_FORMAT_MAP = {
-    "and": "&&",
-    "or": "||",
-}
-
-
-def format_dependency_list(depends, skip_parens=False):
-    assert depends, "Empty dependency list"
-    assert type(depends) == tuple
-
-    if len(depends) == 3:
-        left = format_dependency_list(depends[1])
-        right = format_dependency_list(depends[2])
-
-        operator = OPERATOR_FORMAT_MAP.get(depends[0], depends[0])
-        result = left + " " + operator + " " + right
-        return result if skip_parens else "(" + result + ")"
-    elif depends[0] == 'not':
-        return "!" + format_dependency_list(depends[1])
-    elif depends[0] == 'string':
-        return '"' + depends[1] + '"'
-    elif depends[0] == 'number':
-        return str(depends[1])
-    elif depends[0] == 'boolean':
-        return str(depends[1])
-    elif depends[0] == 'word':
-        return depends[1] + "[=" + str(data.get_config(depends[1])['value']) + "]"
 
 
 def enforce_dependent_values(auto_fix=False):
@@ -398,25 +201,25 @@ def set_initial_values():
         else:
             c['value'] = ''
 
-        for i in dependency_list(c.get("depends")):
+        for i in expr.dependency_list(c.get("depends")):
             data.get_config(i).setdefault('rdepends', []).append(k)
 
         if "default_cond" in c:
             for j in c['default_cond']:
-                for i in dependency_list(j['cond']):
+                for i in expr.dependency_list(j['cond']):
                     data.get_config(i).setdefault('rdefault', []).append(k)
-                value_deps = dependency_list(j['expr'])
+                value_deps = expr.dependency_list(j['expr'])
                 for d in value_deps:
                     data.get_config(d).setdefault('rdefault', []).append(k)
 
         if "default" in c:
-            value_deps = dependency_list(c['default'])
+            value_deps = expr.dependency_list(c['default'])
             for d in value_deps:
                 data.get_config(d).setdefault('rdefault', []).append(k)
 
         if "select_if" in c:
             for j in c['select_if']:
-                for i in dependency_list(j[1]):
+                for i in expr.dependency_list(j[1]):
                     data.get_config(i).setdefault('rselect_if', []).append(k)
 
     # Check for dependency cycles
@@ -457,12 +260,12 @@ def update_choice_default(c):
                 rank = 2
         else:
             for i in config.get("default_cond", []):
-                if can_enable(i['cond']) and expr_value(i['expr']) == 'y':
+                if can_enable(i['cond']) and expr.expr_value(i['expr']) == 'y':
                     rank = 3
                     break
             else:
                 def_expr = config.get("default", expr.NO)
-                if expr_value(def_expr) == "y":
+                if expr.expr_value(def_expr) == "y":
                     rank = 4
 
         return rank, config['position']
@@ -503,12 +306,12 @@ def update_defaults(k):
     if "default_cond" in c:
         for i in c['default_cond']:
             if can_enable(i['cond']):
-                set_config_internal(k, expr_value(i['expr']))
+                set_config_internal(k, expr.expr_value(i['expr']))
                 return
 
     # None of the conditioned defaults match
     if "default" in c:
-        set_config_internal(k, expr_value(c['default']))
+        set_config_internal(k, expr.expr_value(c['default']))
         return
 
     # No default - so set bools to 'n'
@@ -1005,12 +808,12 @@ def menu_parse():
                 continue
 
             while len(depends_stack) > 0:
-                if check_depends(config.get('depends'), depends_stack[-1]):
+                if expr.check_depends(config.get('depends'), depends_stack[-1]):
                     break
                 depends_stack.pop()
 
             while len(menuconfig_stack) > 0:
-                if check_depends(config.get('depends'), menuconfig_stack[-1]):
+                if expr.check_depends(config.get('depends'), menuconfig_stack[-1]):
                     inmenu = menuconfig_stack[-1]
                     break
                 menuconfig_stack.pop()
@@ -1029,7 +832,7 @@ def menu_parse():
             inmenu = config.get('inmenu')
 
             while len(menuconfig_stack) > 0:
-                if check_depends(config.get('depends'), menuconfig_stack[-1]):
+                if expr.check_depends(config.get('depends'), menuconfig_stack[-1]):
                     inmenu = menuconfig_stack[-1]
                     break
                 menuconfig_stack.pop()
